General functions moved to general.cpp
