+++ CMakeLists.txt
@@ -7,7 +7,7 @@ set(CMAKE_CXX_STANDARD 17)
 find_package(Eigen3 REQUIRED)
 find_package(Libint2 REQUIRED)
 
-add_executable(P2 main.cpp hf.cpp mp2.cpp)
+add_executable(P2 main.cpp hf.cpp mp2.cpp general.cpp)
 
 target_link_libraries(P2 Eigen3::Eigen)
 target_link_libraries(P2 Libint2::cxx)
